+++ user-service-smartlogistics/src/main/java/org/dgroup/userservicesmartlogistics/service/DriverService.java
@@ -4,13 +4,15 @@ import org.dgroup.userservicesmartlogistics.model.DriverProfile;
 import org.dgroup.userservicesmartlogistics.model.DriverStatus;
 import org.springframework.security.core.Authentication;
 
+import java.math.BigDecimal;
+
 public interface DriverService {
 
     DriverProfile getDriverProfile(Authentication authentication);
 
     DriverProfile updateDriverLocation(
-            Double latitude,
-            Double longitude,
+            BigDecimal latitude,
+            BigDecimal longitude,
             Authentication authentication
     );
